Tối ưu code 06
