+++ src/layouts/HeaderOnly/Header/Header.js
@@ -1,0 +1,13 @@
+// Logic library import
+
+// Layout library import
+
+// Layout import
+
+// Component import
+
+function Header() {
+    return <h2>Header</h2>;
+}
+
+export default Header;
