+++ src/layouts/HeaderOnly/Header/index.js
@@ -1,13 +1,1 @@
-// Logic library import
-
-// Layout library import
-
-// Layout import
-
-// Component import
-
-function Header() {
-    return <h2>Header</h2>;
-}
-
-export default Header;
+export { default } from './Header';
