+++ src/layouts/HeaderOnly/HeaderOnly.js
@@ -1,0 +1,21 @@
+// Logic library import
+
+// Layout library import
+
+// Layout import
+
+// Component import
+import Header from './Header';
+
+function HeaderOnly({ children }) {
+    return (
+        <div>
+            <Header />
+            <div className="container">
+                <div className="content">{children}</div>
+            </div>
+        </div>
+    );
+}
+
+export default HeaderOnly;
