+++ src/layouts/HeaderOnly/index.js
@@ -1,21 +1,1 @@
-// Logic library import
-
-// Layout library import
-
-// Layout import
-
-// Component import
-import Header from './Header';
-
-function HeaderOnly({ children }) {
-    return (
-        <div>
-            <Header />
-            <div className="container">
-                <div className="content">{children}</div>
-            </div>
-        </div>
-    );
-}
-
-export default HeaderOnly;
+export { default } from './HeaderOnly';
